requirements.txt
asgiref==3.3.4
Django==2.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-redis==4.12.1
djangorestframework==3.12.4
drf-extensions==0.7.0
pytz==2021.1
redis==3.5.3
sqlparse==0.4.1
